applets/mii: Move MiiEdit applet types into its own file
